+++ apps/omsorgspengerutbetaling-soknad/src/app/types/BarnAlderInfo.ts
@@ -1,0 +1,7 @@
+export interface BarnAlderInfo {
+    under13: number;
+    over13: number;
+    totalt: number;
+    kunBarnUnder13: boolean;
+    kunBarnOver13: boolean;
+}