+++ apps/omsorgspengerutbetaling-soknad/src/storybook/decorators/withFormikFormWrapper.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import { StoryFormikFormWrapper } from '../components/StoryFormikFormWrapper';
+
+export const withFormikWrapper = (Story, args) => (
+    <StoryFormikFormWrapper {...args}>
+        <Story />
+    </StoryFormikFormWrapper>
+);