+++ apps/dine-pleiepenger/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @navikt/dine-pleiepenger
 
+## 0.5.16
+
+### Patch Changes
+
+- Pakkeoppdatering patch
+- Updated dependencies
+  - @navikt/sif-common-core-ds@9.15.7
+
+## 0.5.15
+
+### Patch Changes
+
+- Pakkeoppdateringer minor/patch
+- Updated dependencies
+  - @navikt/sif-common-amplitude@2.24.20
+  - @navikt/appstatus-react-ds@2.17.23
+  - @navikt/sif-common-core-ds@9.15.5
+  - @navikt/sif-common-sentry@0.25.18
+
 ## 0.5.14
 
 ### Patch Changes