+++ apps/pleiepenger-sykt-barn/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @navikt/pleiepenger-sykt-barn
 
+## 3.7.31
+
+### Patch Changes
+
+- Pakkeoppdatering patch
+- Updated dependencies
+  - @navikt/sif-common-formik-ds@1.7.16
+  - @navikt/sif-common-soknad-ds@10.13.16
+  - @navikt/sif-common-forms-ds@8.19.5
+  - @navikt/sif-common-core-ds@9.15.7
+
+## 3.7.30
+
+### Patch Changes
+
+- Updated dependencies
+  - @navikt/sif-common-amplitude@2.24.21
+  - @navikt/sif-common-formik-ds@1.7.15
+  - @navikt/sif-common-soknad-ds@10.13.15
+  - @navikt/sif-common-forms-ds@8.19.4
+  - @navikt/appstatus-react-ds@2.17.24
+  - @navikt/sif-common-core-ds@9.15.6
+  - @navikt/sif-common-sentry@0.25.19
+  - @navikt/sif-common-utils@3.44.5
+
+## 3.7.29
+
+### Patch Changes
+
+- Pakkeoppdateringer minor/patch
+- Updated dependencies
+  - @navikt/sif-common-amplitude@2.24.20
+  - @navikt/sif-common-formik-ds@1.7.14
+  - @navikt/sif-common-soknad-ds@10.13.14
+  - @navikt/sif-common-forms-ds@8.19.3
+  - @navikt/appstatus-react-ds@2.17.23
+  - @navikt/sif-common-core-ds@9.15.5
+  - @navikt/sif-common-sentry@0.25.18
+
 ## 3.7.28
 
 ### Patch Changes