+++ packages/sif-common-amplitude/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-amplitude
 
+## 2.24.21
+
+### Patch Changes
+
+- Patch oppdatering av sentry
+
+## 2.24.20
+
+### Patch Changes
+
+- Pakkeoppdateringer minor/patch
+
 ## 2.24.19
 
 ### Patch Changes