+++ packages/sif-common-formik-ds/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @navikt/sif-common-formik-ds
 
+## 1.7.16
+
+### Patch Changes
+
+- Pakkeoppdatering patch
+
+## 1.7.15
+
+### Patch Changes
+
+- Patch oppdatering av sentry
+
+## 1.7.14
+
+### Patch Changes
+
+- Pakkeoppdateringer minor/patch
+
 ## 1.7.13
 
 ### Patch Changes