+++ packages/sif-common-forms-ds/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @navikt/sif-common-forms-ds
 
+## 8.19.5
+
+### Patch Changes
+
+- Pakkeoppdatering patch
+
+## 8.19.4
+
+### Patch Changes
+
+- Patch oppdatering av sentry
+
+## 8.19.3
+
+### Patch Changes
+
+- Pakkeoppdateringer minor/patch
+
 ## 8.19.2
 
 ### Patch Changes