+++ packages/sif-common-sentry/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-sentry
 
+## 0.25.19
+
+### Patch Changes
+
+- Patch oppdatering av sentry
+
+## 0.25.18
+
+### Patch Changes
+
+- Pakkeoppdateringer minor/patch
+
 ## 0.25.17
 
 ### Patch Changes