+++ packages/sif-common-ui/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @navikt/sif-common-ui
 
+## 0.6.11
+
+### Patch Changes
+
+- Pakkeoppdatering patch
+- Updated dependencies
+  - @navikt/sif-common-formik-ds@1.7.16
+  - @navikt/sif-common-core-ds@9.15.7
+
+## 0.6.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @navikt/sif-common-formik-ds@1.7.15
+  - @navikt/sif-common-core-ds@9.15.6
+  - @navikt/sif-common-utils@3.44.5
+
+## 0.6.9
+
+### Patch Changes
+
+- Pakkeoppdateringer minor/patch
+- Updated dependencies
+  - @navikt/sif-common-formik-ds@1.7.14
+  - @navikt/sif-common-core-ds@9.15.5
+
 ## 0.6.8
 
 ### Patch Changes